refactor: code cleanup
1. Group colors into SpotSaverColors for better accessibility.
2. Rename Application theme name to SpotSaverTheme.
3. General code refactor.
