+++ androidApp/src/main/java/com/spot/saver/android/SpotSaverTheme.kt
@@ -8,21 +8,20 @@ import androidx.compose.material3.Typography
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.spot.saver.android.theme.DarkPrimaryColor
-import com.spot.saver.android.theme.DarkSecondaryColor
-import com.spot.saver.android.theme.DarkTertiaryColor
-import com.spot.saver.android.theme.LightPrimaryColor
-import com.spot.saver.android.theme.LightSecondaryColor
-import com.spot.saver.android.theme.LightTertiaryColor
+import com.spot.saver.android.theme.SpotSaverColors.DarkPrimaryColor
+import com.spot.saver.android.theme.SpotSaverColors.DarkSecondaryColor
+import com.spot.saver.android.theme.SpotSaverColors.DarkTertiaryColor
+import com.spot.saver.android.theme.SpotSaverColors.LightPrimaryColor
+import com.spot.saver.android.theme.SpotSaverColors.LightSecondaryColor
+import com.spot.saver.android.theme.SpotSaverColors.LightTertiaryColor
 
 @Composable
-fun MyApplicationTheme(
+fun SpotSaverTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable () -> Unit
 ) {
