+++ androidApp/src/main/java/com/spot/saver/android/theme/SpotSaverColors.kt
@@ -1,0 +1,15 @@
+package com.spot.saver.android.theme
+
+import androidx.compose.ui.graphics.Color
+
+object SpotSaverColors  {
+    val ActionColor = Color(0xFF07527D)
+
+    val DarkPrimaryColor = Color(0xFFBB86FC)
+    val DarkSecondaryColor = Color(0xFF03DAC5)
+    val DarkTertiaryColor = Color(0xFF3700B3)
+
+    val LightPrimaryColor = Color(0xFF6200EE)
+    val LightSecondaryColor = Color(0xFF03DAC5)
+    val LightTertiaryColor = Color(0xFF3700B3)
+}
